Move PlatformRegistry to PackageLoading module
