+++ Sources/PackageLoading/CMakeLists.txt
@@ -14,6 +14,7 @@ add_library(PackageLoading
   ModuleMapGenerator.swift
   PackageBuilder.swift
   PackageDescription4Loader.swift
+  PlatformRegistry.swift
   Target+PkgConfig.swift
   TargetSourcesBuilder.swift
   ToolsVersionLoader.swift
